Revert "Changed x to a UINT16"
This reverts commit d12a51c957936b52f49337620ebbb67a152f3b82.

+++ data/pool_data.h
@@ -5,7 +5,7 @@
 
 typedef struct {
     UBYTE sprite_id[4];
-    UINT16 x; // in tiles so it doesn't overflow
+    UINT8 x; // in tiles so it doesn't overflow
     UINT8 y; // in tiles so it doesn't overflow
     UINT8 sprite_size;
     UINT8 movement_type;
